Created specific "@connection" zone for connection logic
