+++ src/pages/room/Modals/AddSongModal/index.tsx
@@ -2,18 +2,19 @@ import React, { useState } from "react";
 
 import { Modal } from "@components/Modals/Modal";
 import { Button } from "@components/Button";
+import { useEmitRoomData } from "@connection/EmitContext";
+import { useRoomContext } from "@connection/RoomContext";
 import { allSongs } from "@shared/songs";
 
-import { useEmitRoomData } from "../../DynamicScene/RoomConnection/data";
-import { useRoomContext } from "../../RoomContext";
-import { RoomModalProps } from "..";
+import { RoomModalProps } from "../types";
 import { SongSelector } from "./SongSelector";
 import styles from "./styles.module.scss";
 
 export const AddSongModal: React.FC<RoomModalProps> = ({ isOpen, onClose }) => {
-  const [selected, setSelected] = useState<number>();
   const emitRoomData = useEmitRoomData();
   const { roomData } = useRoomContext();
+  const [selected, setSelected] = useState<number>();
+
   const chooseSong = () => {
     if (!selected) {
       return;
