+++ src/pages/room/Modals/index.ts
@@ -2,7 +2,6 @@ import { AddSongModal } from "./AddSongModal";
 import { FeedbackModal } from "./FeedbackModal";
 import { InviteFriendsModal } from "./InviteFriendsModal";
 import { InRoomModal } from "./InRoomModal";
-import { UseModalComponentProps } from "@components/Modals/types";
 
 export const roomModals = {
   addSong: AddSongModal,
@@ -10,9 +9,3 @@ export const roomModals = {
   inRoom: InRoomModal,
   inviteFriends: InviteFriendsModal,
 };
-
-export type RoomModal = "addSong" | "feedback" | "inRoom" | "inviteFriends";
-
-export type RoomModalProps = UseModalComponentProps<RoomModal>;
-
-export type SetRoomModal = (modal: RoomModal) => void;
