+++ src/pages/room/RoomContainer.tsx
@@ -4,15 +4,16 @@ import ReactModal from "react-modal";
 import { rootElementId, modalsElementId } from "@components/constants";
 import { BaseHead } from "@components/meta/BaseHead";
 import { useModals } from "@components/Modals/useModals";
+import { DynamicScene } from "@connection/DynamicScene";
+import { RoomQuery } from "@shared/types";
 
-import { DynamicScene } from "./DynamicScene";
 import { roomModals } from "./Modals";
 import { RoomBottom } from "./RoomBottom";
-import { RoomContextSettings } from "./RoomContext";
+import { RoomEvents } from "./RoomEvents";
 import { RoomHead } from "./RoomHead";
 
 export type RoomContainerProps = {
-  settings: RoomContextSettings;
+  settings: RoomQuery;
 };
 
 export const RoomContainer: React.FC<RoomContainerProps> = ({ settings }) => {
@@ -31,6 +32,7 @@ export const RoomContainer: React.FC<RoomContainerProps> = ({ settings }) => {
       <RoomHead setModal={setModal} />
       <RoomBottom setModal={setModal} />
       {modalNodes}
+      <RoomEvents />
     </DynamicScene>
   );
 };
