+++ src/pages/room/RoomEvents/index.ts
@@ -3,7 +3,7 @@ import React from "react";
 import { useEnvironmentSwitchOnSpace } from "./useEnvironmentSwitchOnSpace";
 import { useVideoControls } from "./useVideoControls";
 
-export const DynamicSceneEvents: React.FC = () => {
+export const RoomEvents: React.FC = () => {
   useEnvironmentSwitchOnSpace();
   useVideoControls();
