+++ tsconfig.json
@@ -15,6 +15,7 @@
     "noUnusedParameters": true,
     "paths": {
       "@components/*": ["components/*"],
+      "@connection/*": ["connection/*"],
       "@shared/*": ["shared/*"]
     },
     "resolveJsonModule": true,
